Split ceilf and time64
