+++ cmake/shared.cmake
@@ -25,7 +25,7 @@ if(LINUX)
         -Wl,--build-id=0x${BUILD_ID}
         ${LIBDDWAF_PRIVATE_LIBRARIES}
         -static-libstdc++
-        glibc_compat)
+        glibc_compat_time64 glibc_compat_math)
 
     if(NOT (CMAKE_BUILD_TYPE MATCHES Debug))
         set(SYMBOL_FILE $<TARGET_FILE:libddwaf_shared>.debug)
@@ -79,5 +79,5 @@ elseif (MINGW)
         -Wl,--build-id=0x${BUILD_ID}
         ${LIBDDWAF_PRIVATE_LIBRARIES}
         -static-libstdc++
-        glibc_compat)
+        glibc_compat_time64 glibc_compat_math)
 endif()
